add orderService logic

+++ Warehouse.BusinessLogicLayer/Services/CartService.cs
@@ -72,6 +72,10 @@ namespace Warehouse.BusinessLogicLayer.Services
         public async Task DeleteCartProductAsync(int CartProductId, ClaimsPrincipal User, string userId = null)
         {
             Cart cp = await _getCartAsync(User, userId);
+            if (cp.CartProducts.Find(c => c.Id == CartProductId) == null)
+            {
+                throw new NotFoundException();
+            }
             await _cartProductRepo.DeleteCartProductAsync(cp.CartProducts.Find(c => c.Id == CartProductId));
         }
     }
